pyproject.toml
[project]
name = "bookit"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "esarr", email = "emmanuelbenoitsarr@gmail.com"},
]
dependencies = [
    "django<4",
    "djongo>=1.3.7",
    "djangorestframework>=3.12,<3.15",
    "drf-yasg>=1.21.8",
    "django-cors-headers>=3.11.0",
    "ruff>=0.8.1",
]
requires-python = ">=3.9, <3.11"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
